Settlement refactor (#11)
* Updated Settlement building placement

rectangles now say what the building is, shows a health bar/percentage, and upgraded the placement system to better match the dimensions on the grid.

* Updated README

* RTS Controller Refactor

Much improved RTS unit control, with group formation, improved selection, camera panning to selected groups

* Final RTS controller refactor

Control Groups (Ctrl+1, 1) are in and working.

Click-Drag Formations are in and working (and facing the right way!).